fix: error when PUBLIC_URL undefined
Only do things post-build when PUBLIC_URL is defined

+++ postbuild.js
@@ -9,15 +9,19 @@
 const fs = require('fs-extra');
 const dotenv = require('dotenv');
 
-// get the tenant name from PUBLIC_URL in .env
-dotenv.config();
-const TENANT_NAME = process.env.PUBLIC_URL.slice(1).split('/')[0]; // get everything between first and second '/'
+if (process.env.PUBLIC_URL) {
+  if (process.env.PUBLIC_URL[0] === '/') {
+    // get the tenant name from PUBLIC_URL in .env
+    dotenv.config();
+    const TENANT_NAME = process.env.PUBLIC_URL.slice(1).split('/')[0]; // get everything between first and second '/'
 
-// copy build folder contents to folder with tenant name
-fs.copySync('./build', `./${TENANT_NAME}`);
+    // copy build folder contents to folder with tenant name
+    fs.copySync('./build', `./${TENANT_NAME}`);
 
-// empty build folder
-fs.emptyDirSync('./build');
+    // empty build folder
+    fs.emptyDirSync('./build');
 
-// move folder with tenant name into empty build folder
-fs.moveSync(`./${TENANT_NAME}`, `./build/${TENANT_NAME}`);
+    // move folder with tenant name into empty build folder
+    fs.moveSync(`./${TENANT_NAME}`, `./build/${TENANT_NAME}`);
+  }
+}
